requirements.txt
django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2
mysqlclient==2.2.7
psycopg2-binary==2.9.11
django-redis==6.0.0
pycryptodomex==3.23.0
djangorestframework-simplejwt==5.5.1
celery==5.5.3
django-celery-beat==2.8.1
django-celery-results==2.6.0
flower==2.0.1
python-daemon==3.1.2
gunicorn==23.0.0
django-proxy==1.3.0
psutil==7.1.1
uvicorn==0.38.0
daphne==4.2.1
channels==4.3.1
channels-redis==4.3.0
django-ranged-response==0.2.0
user-agents==2.2.0
aiofiles==24.1.0
websockets==15.0.1
django-imagekit==6.0.0
pilkit==3.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
openpyxl==3.2.0b1
pyzipper==0.3.6
unicodecsv==0.14.1
chardet==5.2.0
pyexcel==0.7.3
pyexcel-xlsx==0.6.1
alibabacloud-dysmsapi20170525==4.2.0
phonenumbers==9.0.16
pycountry==24.6.1
geoip2==5.1.0
ipip-ipdb==1.6.1
requests==2.32.5
html2text==2025.4.15
pyotp==2.9.0
